+++ llm_ffw/rules/__init__.py
@@ -4,6 +4,7 @@ from .base import Rule, RuleMatch
 from .banned_substrings import BannedSubstringsRule
 from .invisible_characters import InvisibleCharactersRule
 from .json_output import JSONOutputRule
+from .payment_card import PaymentCardRule
 from .secrets import SecretsRule
 from .unsafe_url import UnsafeURLRule
 
@@ -11,6 +12,7 @@ __all__ = [
     "BannedSubstringsRule",
     "InvisibleCharactersRule",
     "JSONOutputRule",
+    "PaymentCardRule",
     "Rule",
     "RuleMatch",
     "SecretsRule",